Version 1.1
Login/Logout script is fully working, links to the dashboard page
Dashboard page contains a place to add an EIR and Faculty member that is
fully functional. You can add information about an EIR/Faculty member
and attach an image (JPG/PNG/GIF with max file size 4MB). Keywords are
separated by comma and whitespace is stripped before inserting into
database.
Search on the homepage has been changed so that the delimiter for
keywords is comma, not space.
Contact form - changed to-send email
Changed ajax php code such that it now requires to connect to db.php
Added JQuery code to show overlay for Faculty members

Things to do:
Have a separate section for EIR and Faculty members when searching on
home page (create different divs)
Have a way to update information on EIR and Faculty members on the
dashboard (Including delete members)
Make it required that you have to sign in to view the dashboard.
Fix requirement issues with form for adding an EIR and Faculty member
Fix display confirmation upon adding a member
Fix display confirmation when entering a wrong username/password in
login screen